collection view code cleanup
