+++ free-tunes/Controllers/OnBoardingViewController.swift
@@ -9,7 +9,10 @@ import UIKit
 
 class OnBoardingViewController: UIViewController {
 
+//  MARK: - Interface Builder Outlets
     @IBOutlet weak var headerText: UILabel!
+   
+//  MARK: - Runtime Methods
     override func viewDidLoad() {
         super.viewDidLoad()
         headerText.buildCustomLabel(labelFirstPart: "free", labelSecondPart: "Tunes")
