+++ free-tunes/Controllers/TrackExplorerViewController.swift
@@ -9,12 +9,15 @@ import UIKit
 import AVFoundation
 
 class TrackExplorerViewController: UIViewController {
-
+    
+//  MARK: - Interface Builder Outlets
     @IBOutlet weak var trackCollectionView: UICollectionView!
+    
     private lazy var exploreViewModel = TrackExplorerViewModel()
     var looper: AVPlayerLooper!
     var player: AVQueuePlayer!
-    
+
+//  MARK: - Runtime Methods
     override func viewDidLoad() {
         super.viewDidLoad()
         trackCollectionView.register(UINib(nibName: "TrackCollectionViewCell", bundle: .main), forCellWithReuseIdentifier: "TrackCollectionViewCell")
@@ -22,14 +25,15 @@ class TrackExplorerViewController: UIViewController {
         getExploreViewContent()
     }
     
-    func setUpExploreCollectionView() {
-        trackCollectionView.dataSource = self
-        trackCollectionView.delegate = self
-    }
-    
     override func viewWillAppear(_ animated: Bool) {
         self.navigationController?.setNavigationBarHidden(true, animated: animated)
         super.viewWillAppear(animated)
+    }
+    
+//  MARK: - Controller Logic Methods
+    func setUpExploreCollectionView() {
+        trackCollectionView.dataSource = self
+        trackCollectionView.delegate = self
     }
     
     func getExploreViewContent() {
@@ -49,8 +53,10 @@ class TrackExplorerViewController: UIViewController {
 
 }
 
+//  MARK: - Collection View Extension
 extension TrackExplorerViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
     
+//  MARK: - Collection View Setup Methods
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         exploreViewModel.numberOfTracks
     }
@@ -67,12 +73,19 @@ extension TrackExplorerViewController: UICollectionViewDelegate, UICollectionVie
         return cell
     }
     
+//  MARK: - Collection View Layout Methods
     func collectionView(_ collectionView: UICollectionView,
                         layout collectionViewLayout: UICollectionViewLayout,
                         sizeForItemAt indexPath: IndexPath) -> CGSize {
       return CGSize(width: view.frame.width, height: view.bounds.height)
     }
     
+    override func viewWillLayoutSubviews() {
+        super.viewWillLayoutSubviews()
+        trackCollectionView.collectionViewLayout.invalidateLayout()
+    }
+    
+//  MARK: - Music Player Extension Methods
     func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
         looper = nil
         loadRadio(radioURL: exploreViewModel.getTrack(atIndex: indexPath.item).trackLink)
@@ -86,11 +99,7 @@ extension TrackExplorerViewController: UICollectionViewDelegate, UICollectionVie
         player.volume = 1.0
     }
     
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-        trackCollectionView.collectionViewLayout.invalidateLayout()
-    }
-    
+//  MARK: - Collection View Custom Methods
     func setupTrackCell(cell: TrackCollectionViewCell?, index: Int ) {
         guard let currentCell = cell, let albumCoverUrl = URL(string: exploreViewModel.getTrack(atIndex: index).album.cover) else {
             return
