+++ free-tunes/ViewModels/TrackExplorerViewModel.swift
@@ -8,16 +8,21 @@
 import Foundation
 
 struct TrackExplorerViewModel {
+    
     private var topTracks: [Track]
+    
     init() {
         self.topTracks = []
     }
+    
     var numberOfTracks: Int {
         return topTracks.count
     }
+    
     func getTrack(atIndex: Int) -> Track {
         return topTracks[atIndex]
     }
+    
     mutating func storeTracksData(tracks: [Track]) {
         topTracks = tracks
     }
